+++ modular_mc/test/packs/data/modular_mc/_shared/_resource_map.ts
@@ -1,0 +1,5 @@
+import { fromFileUrl } from "jsr:@std/path";
+
+export default {
+	sharedBehavior: fromFileUrl(new URL("./shared.behavior.json", import.meta.url)),
+};